Refactor: classroom, reservation 패키지 타입 수정

+++ src/main/java/com/groomiz/billage/classroom/dto/request/AdminClassroomDetailRequest.java
@@ -22,7 +22,7 @@ public class AdminClassroomDetailRequest {
 	private Long building;
 
 	@Schema(description = "층", example = "2")
-	private Integer floor;
+	private Long floor;
 
 	@Schema(description = "강의실 ID", example = "2")
 	private Long classroom;
